+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## [0.2.0] - 2021-02-26
+- added abstract isUsable methode to AbstractPdfCreator
+- added MissingDependenciesException
+- made type property protected in PdfCreatorFactory
+
 ## [0.1.0] - 2021-02-26
 
 Initial release